\name{print.summary.WEE.linear}
\alias{print.summary.WEE.linear}
\title{
Print WEE Linear Summary Object
}
\description{
Print summary of WEE linear regression object
}
\usage{
\method{print}{summary.WEE.linear}(x, ...)
}
\arguments{
  \item{x}{
An object of class "summary.WEE.linear" produced by a call to summary.WEE.quantile()
}
  \item{...}{Optional arguments passed to printing function}
}

\seealso{
\code{\link{summary.WEE.linear}}
}

