\name{print.summary.WEE.logistic}
\alias{print.summary.WEE.logistic}
\title{
Print WEE logistic Summary Object
}
\description{
Print summary of WEE logistic regression object
}
\usage{
\method{print}{summary.WEE.logistic}(x, ...)
}
\arguments{
  \item{x}{
An object of class "summary.WEE.logistic" produced by a call to summary.WEE.quantile()
}
  \item{...}{Optional arguments passed to printing function}
}

\seealso{
\code{\link{summary.WEE.logistic}}
}

