\name{print.summary.WEE.quantile}
\alias{print.summary.WEE.quantile}
\title{
Print WEE quantile Summary Object
}
\description{
Print summary of WEE quantile regression object
}
\usage{
\method{print}{summary.WEE.quantile}(x, ...)
}
\arguments{
  \item{x}{
An object of class "summary.WEE.quantile" produced by a call to summary.WEE.quantile()
}
  \item{...}{Optional arguments passed to printing function}
}

\seealso{
\code{\link{summary.WEE.quantile}}
}
