\name{print.WEE.linear}
\alias{print.WEE.linear}
\title{
Print a WEE.linear object
}
\description{
Print an object generated by WEE.linear
}
\usage{
\method{print}{WEE.linear}(x, ...)
}
\arguments{
  \item{x}{
Object returned from WEE.linear representing the fit of the model
}
  \item{...}{Optional arguments passed to printing function}
}
\seealso{
\code{\link{WEE.linear}}
}
