\name{print.WEE.logistic}
\alias{print.WEE.logistic}
\title{
Print a WEE.linear object
}
\description{
Print an object generated by WEE.logistic
}
\usage{
\method{print}{WEE.logistic}(x, ...)
}
\arguments{
  \item{x}{
Object returned from WEE.logistic representing the fit of the model
}
  \item{...}{Optional arguments passed to printing function}
}
\seealso{
\code{\link{WEE.logistic}}
}
