\name{print.WEE.quantile}
\alias{print.WEE.quantile}
\title{
Print a WEE.linar object
}
\description{
Print an object generated by WEE.quantile
}
\usage{
\method{print}{WEE.quantile}(x, ...)
}
\arguments{
  \item{x}{
Object returned from WEE.quantile representing the fit of the model
}
  \item{...}{Optional arguments passed to printing function}
}
\seealso{
\code{\link{WEE.quantile}}
}
